requirements-dev.txt
flake8
pytest
pytest-sugar
ipython
pydotplus
pytest-django
black
rope
django-extensions==3.1.0
django-debug-toolbar
Werkzeug==1.0.1
